requirements.txt
pykrx>=1.0.45
finance-datareader>=0.9.50
requests>=2.28.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
